Improved warning/detection of incorrect container variables
Summary:
Variables in containers and fragments are a common source of confusion. Relay
currently provides warning when a RelayContainer is rendered without its
fragment having been fetched by the parent component. However, this check does
not take variables into account, making it difficult to debug when the variables
used in a fragment - `Child.getFragment('foo', {size: 32})` - does not match the
variables passed as props: `<Child ... size={64} />`.

This diff refines `RelayFragmentPointer` to not only record which concrete
fragments where fetched for an id, but also which sets of variables were
fetched. `RelayContainer` uses this to do a more accurate check that the correct
fragment + variables were fetched, and gives a much more informative error
message if they weren't.

Reviewed By: wincent

Differential Revision: D3296974

fbshipit-source-id: 03bcdf494457eb57c693e25c3d232364d75796c0