Expose {apply,commit}Update on this.props.relay
Summary:
Provide a way for people who are rendering with a custom environment to dispatch mutations via that environment.

`Relay.Store.{apply,commit}Update` continue to exist, but I have de-emphasized them in the docs and the examples, because more and more people are going to be dealing with custom environments in the future (seeing as we keep recommending them for various use cases).

Closes https://github.com/facebook/relay/issues/1136

Reviewed By: josephsavona

Differential Revision: D3299030

fbshipit-source-id: 8293c8e77bcbd8194244d3703cfc3318ee6b59c4