Relay: fix false positive RelayContainer variable warning
Summary:
The new warning for mismatched query/render variables in RelayContainer doesn't
take into account new variables from calls to `setVariables`. The updated
container variables are compared to the original variables used to fetch the
component which doesn't match and incorrectly triggers the warning.

The fix here is to track the last variables used to render and also compare
next/prev variables:
- On first render the query/render variables are compared as before, there are
  no previous vars to compare against.
- In componentWillReceiveProps, the local variables must match *either* the
  variables used by the parent to fetch the fragment *or* the most recent
  variables used by the component.

Reviewed By: steveluscher

Differential Revision: D3348845

fbshipit-source-id: 359285cc1663ff1bd6b33d5e2211b80a5e27b1e3