Upgrade fbjs preset
Summary:
This removes a bunch of overlapping dependencies now that we do
everything you need in the fbjs preset.

This is pending the actual release of v2 of the preset, which hasn't happened yet - I wanted to try this and see what the outcome would be. Luckily nothing really needed to be done :)

Here's the diff of `lib/`: https://gist.github.com/zpao/79879855ac21d12094087a754daa7fa4. Overwhelmingly the difference is that have the function name transform on now. Otherwise there are a couple ordering differences with the helpers and then L1620 (not sure what's happening there)

It looks like you weren't actually making use of anything that was in stage-1, and es2015 wasn't even used, so got rid of those.
Closes https://github.com/facebook/relay/pull/1139

Differential Revision: D3357595

Pulled By: steveluscher

fbshipit-source-id: e4f29c56e98da0870cdb0339a6ae1f938c2431ee