Unbreak schema updates for graphql 0.6.0
Summary:
This is a partial unbreaking of schema updates given changes to graphql 0.6.0.

"Partial" because for some unknown reason local to my setup, `npm run update-schema` dies unless I create this `.babelrc` file in the root of the project. Will need to dig in further to solve that one.

```
{
  "presets": ["babel-preset-fbjs"]
}
```
Closes https://github.com/facebook/relay/pull/1176

Reviewed By: josephsavona

Differential Revision: D3359035

Pulled By: wincent

fbshipit-source-id: 21bca2fedf2aef946387866b8410989023c01827