Link to the release versions of Relay and the plugin in example code
Summary:
A while ago, we decided to use file: relative version specs in our examples. The idea was that you might want to develop the examples in tandem with Relay or the Babel plugin. Or that you might want to make a backward incompatible change to one of the examples.

In practice, this has kept a lot of people from being able to use the examples because of issues like npm/npm#10379. I'm going to change the dynamic today.

Now, we're going to lock the examples to the release version of react-relay and babel-relay-plugin. This will make the examples run reliably for most of us, and will fix all of the Node 5/6 issues I've just discovered as part of this bug report.

If you want to develop on the examples and Relay in tandem, then you can do a little bit of extra work to npm link your local copy of react-relay and the babel-relay-plugin. A guide on how to set all of that up can be found here: http://justjs.com/posts/npm-link-developing-your-own-npm-modules-without-tears

Closes #1177.
Closes https://github.com/facebook/relay/pull/1183

Differential Revision: D3362071

Pulled By: steveluscher

fbshipit-source-id: a009e8f555db42bc006fa5d58690b11d24528d08