Remove some more unnecessary indirection working with Deferred instances
We don't need to call `getPromise()` on `Deferred` in order to invoke `catch()`;
we can just invoke `catch()` directly.

I did some of this a while back in D3183390, but since then, [[
https://github.com/facebook/fbjs/commit/e3cd4970f5 | e3cd4970f5 ]] landed in
fbjs, adding a `catch` method to `Deferred`, and as of [[
https://github.com/facebook/relay/commit/d653f0e5443 | d653f0e5443 ]] we're
using the latest release of fbjs in Relay, so we can make use of it.