Example to fetch schema from remote GraphQL server
Summary:
I added a simple fully working example for downloading a GraphQL schema from the server.
The examples in the repository only mention the generation of schemas locally.
I struggled a little till I found all the details on how to create a schema json that the babel plugin needs via the Introspection Query, and it also needed a good amount of time to realize a introspection query is in the utilities.

So I thought this might make it easier for other people who do not start of an example, but want to contact a existing GraphQL server.

Have found #1149 recently that is highly related, I just wanted the example to be completely working and still simple, rather than only be an idea people would have to write a script around.
Closes https://github.com/facebook/relay/pull/1199

Reviewed By: kassens

Differential Revision: D3386010

Pulled By: wincent

fbshipit-source-id: 26f0a468901964e15fa8773d8f8c829c8c25067e