Print subscriptions
Summary:
I am not sure that we want to have this support but I think it would be nice.

I replaced the `printMutation` function with a `printOperation` function that can print any instance of `RelayQuery.Operation`. I am not sure if that's the behavior we want or if we should check for `RelayQuery.Subscription` and `RelayQuery.Mutation`.

Also as it is right now we always have to supply a `$input` argument which may not be the case when using subscriptions.
Closes https://github.com/facebook/relay/pull/1223

Reviewed By: josephsavona

Differential Revision: D3463520

Pulled By: wincent

fbshipit-source-id: e4294e4f578f7739902d5c295543178f030a550e