Remove examples
Summary:
We'll be moving these into another repo soon:

https://github.com/relayjs/relay-examples

At the time I write this, that repo contains an exact copy of the examples that I am removing in this commit, but it will be a few days(?) before I can actually flip the switch and make it public.

As such, in this commit I am removing the examples and updating all references to them to instead point to the last commit at which the examples existed. I will follow-up afterwards and update the links to point to the new repo.

Q: Why are we doing it in this order?

A: Because we want to decouple the examples from the main repo in order to make updating them independent of the release process for Relay itself. This will reduce the friction to shipping new versions of Relay and help us to ship more often, starting with the very next release that I am preparing right now.
Closes https://github.com/facebook/relay/pull/1227

Reviewed By: steveluscher, kassens

Differential Revision: D3464551

Pulled By: wincent

fbshipit-source-id: 4f76a51d4af644e354ccaae47ce89952c1d432be