Added pendingVariables to RelayContainer.
Summary:
This is still a WIP, but wanted to create the PR already, to get some eyes on it. Should be able to get it done and tested in our app by the end of the week. Ref #1049

(Note from wincent: applied some fix-ups to this before landing, most important of which is that we keep the `rawVariables` around and store the `pendingVariables` -- which may be transformed via `prepareVariables` -- separately.)

Closes https://github.com/facebook/relay/pull/1069

Reviewed By: josephsavona

Differential Revision: D3358796

Pulled By: wincent

fbshipit-source-id: 317006da6ea281f2d6199daa002bc3132c26f25a