Don't hard-wrap Markdown documents
When preparing the 0.9.1 release notes in `CHANGLOG.md` and previewing them on
GitHub, I noticed some awkward wrapping in the rendered HTML near places where I
had put hard line breaks in the file. This was fixed by not wrapping the 0.9.1
chunk of the release notes.

In this commit I've gone back and unwrapped all the Markdown files where we were
hard-wrapping (the vast majority already don't hardwrap).

I was too lazy to visually inspect all the files, so I identified the likely
candidates with this snippet of Ruby:

```
Dir['docs/*'].map{|d|File.read(d).lines.max_by{|l|l.size}}.map(&:size)
```

Which returned a list of the longest line of all files in "docs/":

```
=> [518, 448, 170, 423, 310, 181, 208, 204, 269, 81, 98, 79, 80, 112, 404,
467, 557, 269, 312, 325, 391, 177, 224, 172, 339, 186, 871, 306, 767, 401,
211, 153]
```

The obvious candidates were the chunk in the middle with lengths 81, 98, 79, 80
and 112, corresponding to the GraphQL docs, and sure enough, they were all
hard-wrapped. Unwrapped by mashing `J` mindlessly in Vim.