Documented all `babel-relay-plugin` options.
Summary:
Some of the options that are supported by `getBabelRelayPlugin` are not documented at all. This just makes sure that all options are listed and have a minimal description of what they do.

I am not using the `validator` myself so did not figure out how exactly that would be used. If it is easy to give a minimal example I would love to add this, so we have some more complete documentation.
Closes https://github.com/facebook/relay/pull/1067

Reviewed By: josephsavona

Differential Revision: D3506803

Pulled By: steveluscher

fbshipit-source-id: d22a91275f79242a7358d7eb30388c935ff9e976