If any item in a plural field is missing data, refetch the whole thing. Otherwise, refetch nothing.
Summary: You may have noticed a regression due to which your plural fields, despite their query not having changed, would not be diffed out in subsequent queries. Instead, what we can do is to take an all-or-nothing approach; refetch all data if any record in a plural field appears to be missing data, or fetch nothing at all if every record has all the data it needs.

Reviewed By: josephsavona

Differential Revision: D3511313

fbshipit-source-id: 26ecbae3c46130465d6cf691496c4e78a4be041f