recycleNodesInto: short-circuit on === values
Summary:
Disovered this while prototyping the new Relay core, where all container data is
deeply frozen in development mode (`__DEV__`). The previous implementation would
attempt to write fields into the "next" data even if it was the same object as
the previous, which was uncovered by the Object.freeze call. Instead the
function can return early.

Reviewed By: steveluscher

Differential Revision: D3538080

fbshipit-source-id: 33df1dad1c76a5fac2afdcf7a36fe072d6eca91d