Update playground frame height to suggest line 20 is not the last line
Summary:
At first it was quite difficult for me to understand the first 2 examples in the website homepage because they basically looked the same.

I then realized it was because they both seemed to have 20 lines of code. The frames did not looke like they were scrollable.

Before
<img width="982" alt="screen shot 2016-07-12 at 14 21 03" src="https://cloud.githubusercontent.com/assets/271047/16766771/432c3602-483c-11e6-9d99-0d84dd6b3a7f.png">

After
<img width="1004" alt="screen shot 2016-07-12 at 14 20 54" src="https://cloud.githubusercontent.com/assets/271047/16766776/4bb93658-483c-11e6-939f-8814ad942f3d.png">

I've changed the height into 426. There might be a better magic number to suggest that line 20 is not the last line of the gist.

I am pretty sure this would have helped me a lot when I first had a look at Relay website.
Closes https://github.com/facebook/relay/pull/1278

Differential Revision: D3567974

Pulled By: kassens

fbshipit-source-id: f2529f01abafff1658ef6d39a4a52a3897fa2e4b