RelayDefaultNetworkLayer, throw Error also on non-200 response.
Summary:
Right now the mutation transacion gets its error attribute set differently on server-side non-200 responses  (e.g. 500) compared to 200 responses that include the `errors` attribute in the json. On the latter it is an `Error` object as described in the docs but on the first we still get the normal fetch `Response`. (So https://facebook.github.io/relay/docs/api-reference-relay-store.html#commitupdate-static-method is not completely right)

That also means handling these two different cases in the mutation `onFailure` call requires two very different pieces of logic.

This PR now makes it so that transaction.getError() always returns an `Error`, and if it was a non-200 response it also includes that status code.
Closes https://github.com/facebook/relay/pull/1163

Reviewed By: josephsavona

Differential Revision: D3344888

Pulled By: steveluscher

fbshipit-source-id: f7ff1f30efac382aa90eaf617e36e583f2d919ff