Parse serialized GraphQL range in getRangeMetadata
Summary:
Resolves #1293

Current if you try to de/serialize your RelayEnvironment you run into an error as `getRangeMetadata` tries calling a `GraphQLRange` method on the serialized array. To prevent this we check to see if we have a serialized instance (which is always an array) and parse it if we do.
Closes https://github.com/facebook/relay/pull/1294

Reviewed By: kassens

Differential Revision: D3602040

Pulled By: josephsavona

fbshipit-source-id: 8283fb3e6786902519ffce648903429585d18d5c