Relay: Reduce extraneous container variable warnings (on non-scalar vars)
Summary:
Addresses extraneous instances of the warning about differing fetched/rendered
fragment variables. An example of the warning:

```
Warning: RelayContainer: component `FooComponent` was rendered with variables that differ from the variables used to fetch fragment `viewer`. The fragment was fetched with variables `{"count":10,"foo":[100],}`, but rendered with variables `{"count":20,"foo":[100],}`. This can indicate one of two possibilities:
 - The parent set the correct variables in the query - `FooComponent.getFragment('viewer', {...})` - but did not pass the same variables when rendering the component. Be sure to tell the component what variables to use by passing them as props: `<FooComponent ... count={...} foo={...} />`.
 - You are intentionally passing fake data to this component, in which case ignore this warning.
```

The issue is that the previous `RelayContainerComparators` call incorrectly
viewed non-scalars as non-equal (which makes sense for `shouldComponentUpdate`
but not the warning).

Reviewed By: yuzhi

Differential Revision: D3641777

fbshipit-source-id: 83e25c9a3e2a5919e8ea8559039127e95c9497dc