Expose injectCacheManager on RelayEnvironment
Summary:
`RelayEnvironment` exposes most of the other injection methods for `RelayStoreData`, so this just adds `injectCacheManager` as well. Even though it's not a documented API, adding it to `RelayEnvironment` will make it easier for those who do want to use it while its "experimental" which paves the way for a stable implementation.

I also added a basic test that just verifies the injection works as expected.
Closes https://github.com/facebook/relay/pull/1320

Differential Revision: D3672206

Pulled By: josephsavona

fbshipit-source-id: 65ccff15fa4b648311bd8fd78474b01f0314c8cf