Add `Relay.disableQueryCache` static method
Summary:
This allows for disabling the query cache in `createRelayQuery` and `buildRQL` by calling `Relay.disableQueryCache()`.

Submitted in response to https://github.com/facebook/relay/issues/754#issuecomment-234995746

Because keys used to store queries include params, the query caches can grow to an unbounded size in certain circumstances (like text search input as a param). This allows the the cache to be disabled manually in circumstances where that is problematic, such as server side rendering.

This PR does not disable the `fragmentCache` in `buildRQL`, which I believe only uses fragments and initial variables, and so has a bounded upper limit in terms of size.
Closes https://github.com/facebook/relay/pull/1302

Reviewed By: josephsavona

Differential Revision: D3792224

Pulled By: wincent

fbshipit-source-id: 9efea99b5422cc2f9ba798f140911d90e1c6842e