Do not reset variables if outside variable props of type Object are the same
Summary:
Closes https://github.com/facebook/relay/issues/1357

When new props are received on a `RelayContainer` a verification step is executed with the new props and the current relay variables and if at least one prop variable has a different value than its corresponding relay variable then the current relay variables are reset to the initial state. But the problem was that `===` was used for comparison and this always return a mismatch for props of type `Object` (arrays, objects) resulting in relay variables always being reset to the initial state even if the prop variables are the same.

See github issue for a detailed example.

The fix was to use `areEquals` which does a deep comparison and can match `Object` javascript types.
Closes https://github.com/facebook/relay/pull/1362

Reviewed By: josephsavona

Differential Revision: D3791469

Pulled By: wincent

fbshipit-source-id: d560b965881c2c848ec9da541c96d4606dc4a3b8