babel-relay-plugin: print custom scalars correctly
Summary:
Fixes #1463; currently argument values are printed incorrectly when the argument type is a custom scalar that serializes as an object. This PR changes the plugin to consider any non-builtin scalar type as equivalent to enums and input objects, which are "printed" as variables instead of being inlined.
Closes https://github.com/facebook/relay/pull/1464

Reviewed By: kassens

Differential Revision: D4272048

Pulled By: josephsavona

fbshipit-source-id: 0b85d0843ea5005ea243e865607bd793b910e1b3