Create BabelPluginGraphQL
Summary:
In order to implement the new `Environment` API methods and product-level APIs (new Containers, etc), we need to  information about a fragment's arguments to exist within the fragment itself so that it can be statically analyzed. This is accomplished via an `argumentDefinitions` directive on fragment definitions that will replace `initialVariables`. `Module.getFagment('name')` calls are similarly replaced with `...Module_name`, with an `arguments` directive to replace the second argument of `getFragment`.

There's kind of a circular dependency between code that uses the output of this new plugin and the plugin itself, so the previous diff D4304278 works around it by hard-coding the concrete nodes instead of parsing queries.

Reviewed By: kassens

Differential Revision: D4302558

fbshipit-source-id: 8da8624d24ec5b9e92193e1040dd6586939ef7b5