Add RelaySelector helper class, start of updated Environment interface
Summary:
Defines the `Selector` interface and implements helpers for converting fragment
props into Selectors. This is a building block for implementing
`FragmentSpecResolver` that will be the foundation of the new container APIs.

As part of making the test work I had to fix an issue with `BabelPluginGraphQL`.
It outputs a root fragment for a query, but this conflicted because the test
schema uses the name "Root" for the query type instead of the defacto standard
"Query". This then necessitated a fix to our validation of the `node` field,
which needed to check that the `node` field was not on the query type.

Reviewed By: wincent

Differential Revision: D4309955

fbshipit-source-id: e97c767dcb37974d3359165f011ff0fe63f87304