RelayCore type and legacy implementation
Summary:
Defines the `RelayCore` public API and makes it accessible on the `Environment`
interface via an `unstable_internal: RelayCore` property. Combined with the
direct instance methods of the `Environment` class (which are more stable), these
functions define everything that we have found necessary to implement the new
containers and renderer. Additional methods will be added in follow-ups for
mutations.

Note that this diff only updates the legacy core, a follow-up will modify the
new core to match.

Reviewed By: kassens

Differential Revision: D4334481

fbshipit-source-id: f95b9c6745eb68cebf9eee1563f4cbd6ff353238