ReactRelay2: Access internals via environment.unstable_internal
Summary:
Changes the React/Relay2 layer to access all internals via
`environment.unstable_internal.*` instead of directly requiring the Relay2
versions of functions. Example: previously the renderer directly required
`Relay2OperationSelector` to call `createOperationSelector`, now it uses the
version of that function passed on the environment.

As of this diff, the only blocker for using ReactRelay2* APIs with either the
legacy or new cores is the different structure of the various template tags:
the next diff will unify these.

Note: I removed the `applyMutation` check from `isRelayEnvironment` because this
function isn't ready yet in the new core; everything works without this check
and we can add it back in a few diffs.

Depends on D4348015

Reviewed By: kassens

Differential Revision: D4348810

fbshipit-source-id: ac50e6bd23824a346200e7af610f44b4128d11a4