RelayGraphQLTag: interop with various tag formats
Summary:
The ultimate goal is to unify the output of several Relay tags for GraphQL
template expressions. As a first step, this teaches `RelayGraphQLTag` (and the
Relay2 equivalent) to understand all of the output formats so that we can
change code that accesses the template output first and unify their output later.

Depends on D4348810

Reviewed By: kassens

Differential Revision: D4349583

fbshipit-source-id: 71cbb1b336e80076cb30b51c9571b176e5d18250