Relay: legacy plugin prints types for id/numeric/list args
Summary: Changes the legacy plugin to print `metadata.type` information for field arguments whose type is ID, numeric, or a list since values for these types may need to represented as a string in JavaScript due to its maximum integer size. The presence of this metadata tells the printer to print such arguments as variables instead of inlining them.

Reviewed By: kassens

Differential Revision: D4356786

fbshipit-source-id: f62aa6fde531336b5ad4d36ad60d4091b0856494