Allow injecting default global variables
Summary:
See code comments: this is a workaround for the fact that product code using the
new container API are referencing product-specific global variables, but
existing routes do not specify values for such variables. The simplest place I
could find to inject these was in `RelayQuery`. I'm open to other suggestions,
but this is simple/straightforward and works.

Reviewed By: yuzhi

Differential Revision: D4363616

fbshipit-source-id: 744c5a25bea96c5aafddf867adac1adaeeb619c7