Fix Flow issues related to util core module in plug-in
Mitigates: #1508

Flow is complaining about these `require('util')` calls in the plug-in,
reporting errors like:

```
src/getBabelRelayPlugin.js:21
21: const util = require('util');
  This modules resolves to
  "<<PROJECT_ROOT>>/../../node_modules/util/package.json", which is
  outside both your root directory and all of the entries in the
  [include] section of your .flowconfig. You should either add this
  directory to the [include] section of your .flowconfig, move your
  .flowconfig file higher in the project directory tree, or move this
  package under your Flow root directory.
```

Confirmed with the Flow team that this is a bug in the resolver
implementation. This page indicates that it is supposed to look for core
modules before searching `node_modules`:

https://nodejs.org/api/modules.html#modules_all_together

As a workaround for now, let's just suppress the errors.