Move babel-plugin-graphql back from relay2 -> relay
Summary:
This was in the GitHub repo previously, but it was moved out of it when we did some internal restructuring (internal diff D4370422, GitHub commit d9b607bf4e28aecce1b4dd049). We need to put it back in the GitHub repo, however, if we want to be able to run the test suite in OSS at all (because the test suite has `graphql` template literals throughout).

Note that we already had GitHub <-> FB sync configured for this directory, so I am going to leave it at this path, even though I do think that having it at the top level of the repo is suboptimal.

Reviewed By: leebyron

Differential Revision: D4492763

fbshipit-source-id: 60ecf9772ed019b875bc57686348ca6b6f20b395