Fix require for regenerator-runtime
`package.json` has a dependency on `babel-runtime` which pulls in
regenerator-runtime. In npm 3 we were able to require `regenerator-runtime`
directly due to flattening, but we should only require what we list in the
package.json directly.

Going the route via babel-runtime helps us keep the runtime and compiler in
sync.