Fix require for regenerator-runtime
Summary:
`package.json` has a dependency on `babel-runtime` which pulls in
regenerator-runtime. In npm 3 we were able to require `regenerator-runtime`
directly due to flattening, but we should only require what we list in the
package.json directly.

Going the route via babel-runtime helps us keep the runtime and compiler in
sync.
Closes https://github.com/facebook/relay/pull/1591

Reviewed By: leebyron

Differential Revision: D4848075

Pulled By: kassens

fbshipit-source-id: 8016c5d374f3cbd5819fa7a13dd507b41b513b5b