Disable node 4 on Travis
Summary:
In order to run travis tests on node 4 we would need to either compile more of
the CI infra with babel or remove syntax features that are not supported by
node 4. Node 4 tests also run a lot slower and occationally segfault on travis
for some reason. Let's disable them for the time being and see if there's a
need.
Closes https://github.com/facebook/relay/pull/1595

Reviewed By: leebyron

Differential Revision: D4856506

Pulled By: kassens

fbshipit-source-id: 29aca4aa8549725e56e0d51a1ab659c9457f9722