Warn if incompatible types are encountered for the same id
Summary: Warns (in __DEV__ only) if the same id is assigned different `__typename`s in a payload. This *should* never happen, but it's not at all obvious when it does. Given that this should be rare in practice, this diff adds a check for the condition only in `__DEV__`, avoiding overhead in production.

Reviewed By: kassens

Differential Revision: D4860309

fbshipit-source-id: bd33613824d5379345c26fed39182a3cfd0fa3b2