babel-plugin-relay: make modern the default mode.
Summary:
This replaces the older "relay" and "relayExperimental" keys with "classic" and "modern" respectively and flips the configuration to default to modern.

To ensure cache consistency when pulling new changes while the RN packager is running, I've included fallbacks to the existing keys. We can remove these in a follow-up release.

Reviewed By: kassens

Differential Revision: D4872547

fbshipit-source-id: c3b6388846c1762834834a528e3399d52c2472f4