Add deploy script for making npm deploys easier
Summary:
This adds a bash script which looks for common mistakes before running deploys. It looks to ensure all versions match, that git has pulled, that it's not on a feature branch, and a tag has been applied to the current comment. It also determines if it is a pre-release publish.
Closes https://github.com/facebook/relay/pull/1606

Reviewed By: kassens

Differential Revision: D4880539

Pulled By: leebyron

fbshipit-source-id: 943f25a8e883d18fcf503a9143b0282fe0418408