Fix LeafType mistyping
RelayIR assumes the type of a ScalarField is `GraphQLLeafType`, however that does not account for the possibility of nullable or list variations. When `isLeafType` was updated in graphql@0.9.2 to match the definition of LeafType this mistyping started causing errors.

This proposes typing ScalarField's type as `GraphQLOutputType` which is a little broader than necessary, but importantly includes the possibilities of nullable and list variations. It also explicitly calls getNamedType before passing to isListType to ensure the check is on the unwrapped named type.

Also ups the dependency on graphql to the latest release to illustrate Travis test passing