convert old fixtures to golden test format
Summary:
Converts the fixtures from the old `babel-relay-plugin` to the fixture format we
use everywhere else. In a following change we can run the new plugin on the old
fixtures.

The small node script that generated the conversion is included for illustration
or in case we need to re-run it.

Reviewed By: leebyron

Differential Revision: D4891467

fbshipit-source-id: 5afad6225122db1983c303ff30e858f07b8da381