only write .flow files for whitelist
Summary:
This adds a whitelist for fragment names for which we generate a `.flow.js`
file. Users should be changed to use the flow type exported from the
`.graphql.js` file instead. This will be useful, because I'm planning to change
the flow types in the .graphql files to no longer export the subtypes and switch
them over to use the new `RelayFlowGenerator`.

Reviewed By: leebyron

Differential Revision: D4918307

fbshipit-source-id: 250668189e2a2412ecf9c244029acf4b5c5c105a