Allow non null types in connection validation
Summary:
The current connection validation is too strict and doesn't allow non-null types for a lot of things. It also isn't compatible with the types generated by `graphql-relay-js`.

This uses `RelaySchemaUtils.getNullableType` before instanceof checks to make sure we get the actual type and not `GraphQLNonNull`. Also split some invariants into 2 different checks to keep it readable.

Tested in an app that uses a PaginationController with a schema generated with `graphql-relay-js`.
Closes https://github.com/facebook/relay/pull/1637

Differential Revision: D4917068

Pulled By: alangenfeld

fbshipit-source-id: d4ba50342cbb80c6ff39a6279ea5ac769d062e05