Fix modern containers
Summary:
Fixes an issue where modern containers cannot access any Relay data
(https://github.com/facebook/relay/issues/1665).

The issue is caused by `ReactRelay*Container` (aka modern containers) not
unwrapping the output of the `graphql` tag. Previously this was handled by
CompatContainerBuilder, but this was removed as part of an unrelated fix in
order to not require legacy code in modern mode.

This diff adds a `ReactRelayContainerBuilder` that is a simplified form of
CompatContainerBuilder that only unwraps the results of `graphql` tags, but does
not add any compatibility code.

Note that the idea is for all `ReactRelay*` modules to treat the shape of the
`graphql` type as opaque, retrieving the actual value from the environment's
`getFragment()` and `getOperation()` functions. For containers, this value isn't
known until the container is rendered and the environment is available via
context. The change here makes containers work similarly to `QueryRenderer`,
which also unwraps its `query` value rather than assume a particular shape.

Reviewed By: leebyron

Differential Revision: D4939174

fbshipit-source-id: 1a7f08064441654740e46353bfa47e134142d980