Improve QueryRenderer graphql naming convention
Summary:
I've been playing a little bit with the release candidate and I was stuck on the relay compiler failing to parse the file. After checking the source code of the compiler I realised that it has a strict check on the expected query name.

I'm adding a small paragraph trying my best to explain how it works to avoid some confustion. In the case that someone in Facebook has had the time to improve documentations feel free to close the PR.
Closes https://github.com/facebook/relay/pull/1646

Differential Revision: D4939993

Pulled By: leebyron

fbshipit-source-id: b37fd1d557425e86cc36afdb772079b379d731e5