Allow non null connections
Summary:
`GraphQLNonNull<Connection>` caused relay compiler errors. This adds a few more `getNullableType` to make sure it works with non nullable fields.

Had some issues getting it to pass flow properly and had to change the signature of `getNullableType` because of some incompatible type errors. Since this is in the graphql-js package I just copied it in for now and if the change seems fine I can PR on the graphql-js  repo and remove it when updated.
Closes https://github.com/facebook/relay/pull/1688

Differential Revision: D4965386

Pulled By: leebyron

fbshipit-source-id: f4d71dda14f94094b9e298190d545a15047ce9e9