[RFC] Don't delete invisible files in CodegenDirectory
In a side-project, I have put an `.eslintrc` file in my `__generated__`
directories in order to stop eslint complaining about the use of "'use
strict';" in the generated files. Specifically, it contains:

```
{
  "rules": {
    "strict": [0]
  }
}
```

Running the compiler, however, causes those invisible files to be blown
away:

```
Created:
 - FooQuery.graphql.js
Deleted:
 - .eslintrc
 - .eslintrc
```

This commit makes it so that we never delete invisible files (ie. files
starting with a dot), which will allow people to stick `.eslintrc`,
`.gitignore` etc in their `__generated__` directories without
interference. We might want to consider a whitelist approach instead
(ie. only delete files with a ".graphql.js" extension), but that would
be a more invasive change.