Don't delete invisible files in CodegenDirectory
Summary:
In a side-project, I have put an `.eslintrc` file in my `__generated__` directories in order to stop eslint complaining about the use of `'use strict';` in the generated files. Specifically, it contains:

```
{
  "rules": {
    "strict": [0]
  }
}
```

Running the compiler, however, causes those invisible files to be blown away:

```
Created:
 - FooQuery.graphql.js
Deleted:
 - .eslintrc
 - .eslintrc
```

This commit makes it so that we never delete invisible files (ie. files starting with a dot), which will allow people to stick `.eslintrc`, `.gitignore` etc in their `__generated__` directories without interference. We might want to consider a whitelist approach instead (ie. only delete files with a ".graphql.js" extension), but that would be a more invasive change.
Closes https://github.com/facebook/relay/pull/1743

Reviewed By: mjmahone

Differential Revision: D5043667

Pulled By: wincent

fbshipit-source-id: 7d75d57db20fa7da26808173063a21bf3fe2b778