Fix for polling-based live queries.
Summary:
Currently polling-based live queries don't get any data right away.
Was this intentional? My assumption is that this was considered to be working
fine because it was tested either with a low enough interval such that the
delay just looked like normal loading time, or that it was tested on an app
that had server side initial layout rendering. I think it makes sense as a
default behavior to do a fetch immediately, but this is more of a breaking
change than, say, requiring `cacheConfig={doImmediateFetch: true}`.

Reviewed By: JenniferWang

Differential Revision: D5108459

fbshipit-source-id: d05c47a855f0d03952f3906a9ecbbdf193b18e93