Generate plural Flow types
Previously, if you declared a fragment with the directive
`@relay(plural: true)` the generated Flow type would be an object of the
declared type. It seems this was an oversight, and the resulting type
should actually be an array of that type. This commit checks to see
whether the `@relay` directive was given with a `plural` argument, and
if it was, wraps the object type in an array type.