Fix ReferenceError in mutations docs
I'm _fairly_ sure that the code as previously written would result in a `ReferenceError` due to `source` and `storyId` not having been declared at the top-level scope. I think the intent here was to pull the values for `variables.input` from the `source` and `storyId` params in the `markNotificationAsRead` method, so I moved the `variables` declaration inside that method.

All of that said, my understanding of Relay is very limited, so my apologies if there is some Relay magic going on here I'm just not grokking.