Generate plural Flow types
Summary:
I'm pretty new to Relay/GraphQL, so I wanted to run a failing test by the team to make sure this is the expected behavior before I spend time fixing the actual issue. If somebody gives me the go ahead I can look into fixing this.

Currently, if you declare a fragment with the directive `relay(plural: true)` the generated Flow type is an object of that type. I would expect this to be an array of that type. This commit adds a failing test to reproduce this scenario.
Closes https://github.com/facebook/relay/pull/1810

Reviewed By: JenniferWang

Differential Revision: D5143460

Pulled By: kassens

fbshipit-source-id: d5658f3280b6b9b260c4a180fb23222ae08c8b4a