Teach compiler `--persist` option
Sample usage:

    relay-compiler --src ./src --schema schema.graphql --persist scripts/persistQuery.js

Where `persistQuery.js` looks like this:

```
function persistQuery(text) {
  var id = someLogicToPersistQueryText(text);
  return Promise.resolve(id);
}
```

The alternative in the absence of this option is to copy-paste the
entire `RelayCompilerBin` implementation just to override that bit.

Note that webpack makes this ghastly — necessitating the use of
`eval` — because it wants to rewrite the `require` of the passed
in module, which of course fails because it is dynamic. Google search
turns up a bunch of results like this one lamenting the situation:

https://stackoverflow.com/questions/30575060/require-js-files-dynamically-on-runtime-using-webpack

Apparently it's better in webpack 2.