remove soft error mode
Summary:
This removes a "soft error" mode for the Relay Babel plugin where instead of
failing to transform code, the plugin would generate a `throw new Error(...)`
statement. This turns out to be a footgun since it removes the compiler safety
from guarding agains shipping code that's invalid according to the schema.

Reviewed By: leebyron

Differential Revision: D5234525

fbshipit-source-id: 593e57dc6a95d750dad3f0e99643454ca39c2826