Change `optimisticResponse` to be an object instead of a function
Summary:
The only advantage of making `optimisticResponse` a function is laziness;
however if we actually needs to validate the `optimisticResponse`, which means
the function might be executed multiple times, there is no gauranteed that
the user-provided `optimisticResponse` has no side-effect. Being a function, it could
accidentally capture variables(which might change) in the outer scope like in
`commitRelayModernMutation-test`.

Reviewed By: kassens

Differential Revision: D5253492

fbshipit-source-id: 07be06e49884dee0347cf0452f12992cb1336ad5